Move `ValidationErrors` into `common.results`
